Make `RunnableNode` take `RestApiBuilder`
This change alters the `ActixWeb1` Rest API variant that is passed to a
`RunnableNode` from a `RestApi` struct to a `RestApiBuilder`. This
allows for the authorization configuration of the REST API to be added
to as resources are added.
